+++ source/sink/service_impl.cc
@@ -35,7 +35,7 @@ grpc::Status SinkServiceImpl::StoreExecutionResponseStream(
 grpc::Status abslStatusToGrpcStatus(const absl::Status& status) {
   grpc::Status grpc_status =
       status.ok() ? grpc::Status::OK : grpc::Status(grpc::StatusCode::INTERNAL, status.ToString());
-  ENVOY_LOG_MISC(trace, "Finishing stream with status {} / message {}.", grpc_status.error_code(),
+  ENVOY_LOG_MISC(trace, "Finishing stream with status {} / message {}.", static_cast<int>(grpc_status.error_code()),
                  grpc_status.error_message());
   return grpc_status;
 }
